refactor(plugins): PR-2-2A 纯 API 型 5 插件迁移 + registry 顺序修正 + 核心存储服务容器
- registry.listTools() 改为整体按 tool-order.json 遍历（core+plugin 混合取），插件化后 tools/list 顺序仍逐字节=基线
- start-monitor 新增 registerCoreServices：把 14 个存储方法注册为 storage.<name> 服务供插件 api.consume
- plugin-api 的 api.vrchat 增加 uploadImageFile/uploadPrint/uploadGalleryImage/download 委托（复用 ctx.api，含凭据+限流）
- 迁入 plugins/official/：groups(9)/favorites(7)/media(10)/planet(2)/booth(4)，handler 内 ctx.api→api.vrchat.fetch、ctx.storage→api.consume、外网(booth/planet)→Node fetch
- 删除 core/tools 对应 7 个源文件；search_worlds 摘到 core/tools/misc.js（world-kb 待批次B）；recommend-worlds.js 内联 planet 抓取 + import 改 misc
- check-doc-drift 复用 scripts/dump-tools.mjs(加载插件后抖全部工具名) 作权威源；test-registry 改为插件加载版

验收: 5 插件 loaded；listTools 91 逐字节一致；dispatch 全路由到 handler；safe-mode 过滤插件破坏性工具(remove_print/remove_gallery_image/unfavorite_friend/move_friend_group/leave_group)；check-doc-drift code_tools=91/doc_tools=91 无漂移；test-registry 默认+safe 均 PASS
