+++ core/tools/misc.js
@@ -400,6 +400,41 @@ export async function handleBackupDatabase() {
   }
 }
 
+export async function handleSearchWorlds({ query, n }) {
+  const { api, storage } = ctx;
+  if (!query || typeof query !== 'string') throw new Error('query is required');
+  const limit = Math.min(Math.max(parseInt(n, 10) || 10, 1), 30);
+  const apiWorlds = [];
+  try {
+    const r = await api._request('GET', `/worlds?search=${encodeURIComponent(query)}&n=${limit}`);
+    if (r.status === 200) {
+      for (const w of (r.data || [])) {
+        apiWorlds.push({
+          worldId: w.id,
+          name: w.name,
+          authorName: w.authorName,
+          capacity: w.capacity,
+          imageUrl: w.imageUrl,
+          description: (w.description || '').slice(0, 200),
+        });
+      }
+    }
+  } catch (e) { /* API 失败时仅用本地结果 */ }
+
+  const local = storage.searchWorldsByName(query);
+
+  // 合并：API 结果优先（完整信息），本地补充（可能命中 API 搜不到的）
+  const seen = new Set(apiWorlds.map(w => w.worldId));
+  const merged = [...apiWorlds];
+  for (const lw of local) {
+    if (!seen.has(lw.worldId)) {
+      seen.add(lw.worldId);
+      merged.push({ worldId: lw.worldId, name: lw.name });
+    }
+  }
+  return { query, apiCount: apiWorlds.length, localCount: local.length, count: merged.length, worlds: merged };
+}
+
 // ── MCP 自声明工具表 ──
 export const tools = [
   {
@@ -788,5 +823,26 @@ export const tools = [
       "properties": {}
     },
     handler: async (args) => handleBackupDatabase(args)
+  },
+  {
+    "name": "search_worlds",
+    "description": "[query] Search VRChat worlds by name. English/Japanese search the live API; Chinese keywords fall back to local cache (API CJK search is unreliable).",
+    "inputSchema": {
+      "type": "object",
+      "properties": {
+        "query": {
+          "type": "string",
+          "description": "World name keyword (Chinese/English/Japanese)"
+        },
+        "n": {
+          "type": "number",
+          "description": "Max API results (default 10, max 30)"
+        }
+      },
+      "required": [
+        "query"
+      ]
+    },
+    handler: async (args) => ctx.rateLimiter.execute(() => handleSearchWorlds(args))
   }
 ];
