+++ plugins/official/booth/plugin.json
@@ -1,0 +1,5 @@
+{
+  "name": "booth",
+  "version": "1.0.0",
+  "description": "BOOTH.pm VRChat asset search and cached item tracking"
+}
