+++ plugins/official/favorites/plugin.json
@@ -1,0 +1,5 @@
+{
+  "name": "favorites",
+  "version": "1.0.0",
+  "description": "VRChat 世界/好友收藏管理（云端收藏夹查询、添加、移除、移动）"
+}
