+++ plugins/official/groups/plugin.json
@@ -1,0 +1,5 @@
+{
+  "name": "groups",
+  "version": "1.0.0",
+  "description": "VRChat 群组查询、搜索、加入/退出、公告偷看及热度分析"
+}
