+++ plugins/official/media/plugin.json
@@ -1,0 +1,5 @@
+{
+  "name": "media",
+  "version": "1.0.0",
+  "description": "VRChat media tools: boop emojis, prints, gallery, uploads and downloads"
+}
