+++ plugins/official/planet/plugin.json
@@ -1,0 +1,5 @@
+{
+  "name": "planet",
+  "version": "1.0.0",
+  "description": "PlanetVRC world search and recommendations (planetvrchat.net)"
+}
